// src/app/swaggerFiles/model/evenement.ts
/**
 * assoparenteleve API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Distribution } from './distribution';
import { Produit } from './produit';

export interface Evenement { 
    id?: number;
    nom?: string;
    dateDebut?: string;
    dateFin?: string;
    datePaiement?: string;
    distributionList?: Array<Distribution>;
    produitList?: Array<Produit>;
    valider?: boolean;
    avecEleve?: boolean;
}