// src/app/swaggerFiles/model/produit.ts
/**
 * assoparenteleve API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Produit { 
    id?: number;
    nom?: string;
    fournisseur?: string;
    prix?: number;
    option?: boolean;
}